HBX-1662: Move class 'org.hibernate.tool.hbm2x.pojo.AnnotationBuilder' to 'org.hibernate.tool.internal.util.AnnotationBuilder'
