+++ orm/src/main/java/org/hibernate/tool/internal/util/AnnotationBuilder.java
@@ -1,4 +1,4 @@
-package org.hibernate.tool.hbm2x.pojo;
+package org.hibernate.tool.internal.util;
 
 import java.util.ArrayList;
 import java.util.Iterator;
